+++ src/main/java/hiveudfs/udf/date/DatePatternType.java
@@ -1,0 +1,7 @@
+package hiveudfs.udf.date;
+
+final class DatePatternType {
+    public static final int DATE = 1;
+    public static final int TIMESTAMP = 2;
+    public static final int TIMESTAMP_TZ = 3;
+}
