+++ src/main/java/hiveudfs/udf/date/DateUnitEnum.java
@@ -1,17 +1,6 @@
 package hiveudfs.udf.date;
 
-final class DateUnitType {
-    public static final int YEAR = 1;
-    public static final int MONTH = 2;
-    public static final int DATE = 3;
-    public static final int HOUR = 4;
-    public static final int MINUTE = 5;
-    public static final int SECOND = 6;
-    public static final int QUARTER = 7;
-    private DateUnitType (){}
-}
-
-public enum DateUnit {
+public enum DateUnitEnum {
     YEAR("YEAR", DateUnitType.YEAR),
     YYYY("YYYY", DateUnitType.YEAR),
     YY("YY", DateUnitType.YEAR),
@@ -33,7 +22,7 @@ public enum DateUnit {
     private final String name;
     private final int type;
 
-    DateUnit(String format, int formatType) {
+    DateUnitEnum(String format, int formatType) {
         this.name = format;
         this.type = formatType;
     }
@@ -46,16 +35,16 @@ public enum DateUnit {
     }
 
     public static boolean isMember(String name) {
-        for (DateUnit dateUnit : values()) {
-            if (dateUnit.getName().equals(name)) {
+        for (DateUnitEnum DateUnitEnum : values()) {
+            if (DateUnitEnum.getName().equals(name)) {
                 return true;
             }
         }
         return false;
     }
     public static boolean isMember(String name, int formatType) {
-        for (DateUnit dateUnit : values()) {
-            if (dateUnit.getType() == formatType && dateUnit.getName().equals(name)) {
+        for (DateUnitEnum DateUnitEnum : values()) {
+            if (DateUnitEnum.getType() == formatType && DateUnitEnum.getName().equals(name)) {
                 return true;
             }
         }
