+++ src/main/java/hiveudfs/udf/date/DateUnitType.java
@@ -1,0 +1,12 @@
+package hiveudfs.udf.date;
+
+final class DateUnitType {
+    public static final int YEAR = 1;
+    public static final int MONTH = 2;
+    public static final int DATE = 3;
+    public static final int HOUR = 4;
+    public static final int MINUTE = 5;
+    public static final int SECOND = 6;
+    public static final int QUARTER = 7;
+    private DateUnitType (){}
+}
